refactor(voyager): rename consensus modules to finality modules
