+++ Cargo.toml
@@ -157,14 +157,14 @@ members = [
   "voyager/modules/client-bootstrap/state-lens/ics23-smt",
   "voyager/modules/client-bootstrap/state-lens/ics23-ics23",
 
-  "voyager/modules/consensus/bob",
-  "voyager/modules/consensus/arbitrum",
-  "voyager/modules/consensus/berachain",
-  "voyager/modules/consensus/cometbls",
-  "voyager/modules/consensus/ethereum",
-  "voyager/modules/consensus/movement",
-  "voyager/modules/consensus/tendermint",
-  "voyager/modules/consensus/trusted-evm",
+  "voyager/modules/finality/bob",
+  "voyager/modules/finality/arbitrum",
+  "voyager/modules/finality/berachain",
+  "voyager/modules/finality/cometbls",
+  "voyager/modules/finality/ethereum",
+  "voyager/modules/finality/movement",
+  "voyager/modules/finality/tendermint",
+  "voyager/modules/finality/trusted-evm",
 
   "voyager/plugins/client-update/bob",
   "voyager/plugins/client-update/arbitrum",
