--- voyager/modules/consensus/cometbls/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name    = "voyager-consensus-module-cometbls"
-version = "0.0.0"
-
-authors      = { workspace = true }
-edition      = { workspace = true }
-license-file = { workspace = true }
-publish      = { workspace = true }
-repository   = { workspace = true }
-
-[lints]
-workspace = true
-
-[dependencies]
-cometbft-rpc    = { workspace = true }
-embed-commit    = { workspace = true }
-jsonrpsee       = { workspace = true, features = ["macros", "server", "tracing"] }
-serde           = { workspace = true, features = ["derive"] }
-thiserror       = { workspace = true }
-tokio           = { workspace = true }
-tracing         = { workspace = true }
-unionlabs       = { workspace = true }
-voyager-message = { workspace = true }
-voyager-vm      = { workspace = true }
