--- voyager/modules/consensus/movement/src/main.rs
@@ -1,172 +1,0 @@
-use aptos_rest_client::error::RestError;
-use jsonrpsee::{
-    core::{async_trait, RpcResult},
-    types::ErrorObject,
-    Extensions,
-};
-use serde::{Deserialize, Serialize};
-use tracing::{debug, trace};
-use unionlabs::{
-    aptos::{
-        account::AccountAddress, state_proof::StateProof,
-        transaction_proof::TransactionInfoWithProof,
-    },
-    ibc::core::client::height::Height,
-    primitives::H160,
-    ErrorReporter,
-};
-use voyager_message::{
-    module::{ConsensusModuleInfo, ConsensusModuleServer},
-    primitives::{ChainId, ConsensusType, Timestamp},
-    vm::BoxDynError,
-    ConsensusModule,
-};
-
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-struct StateProofResponse {
-    tx_index: u64,
-    state_proof: StateProof,
-    tx_proof: TransactionInfoWithProof,
-}
-
-#[tokio::main(flavor = "multi_thread")]
-async fn main() {
-    Module::run().await
-}
-
-#[derive(Debug, Clone)]
-pub struct Module {
-    pub chain_id: ChainId,
-
-    /// The address of the IBC smart contract.
-    pub ibc_handler_address: AccountAddress,
-
-    /// The address of the settlement contract on Eth.
-    pub l1_settlement_address: H160,
-
-    pub l1_client_id: u32,
-
-    pub aptos_client: aptos_rest_client::Client,
-
-    pub movement_rest_url: String,
-}
-
-impl ConsensusModule for Module {
-    type Config = Config;
-
-    async fn new(config: Self::Config, info: ConsensusModuleInfo) -> Result<Self, BoxDynError> {
-        let aptos_client = aptos_rest_client::Client::new(config.aptos_rest_api.parse().unwrap());
-
-        let chain_id = aptos_client.get_index().await?.inner().chain_id;
-
-        info.ensure_chain_id(chain_id.to_string())?;
-        info.ensure_consensus_type(ConsensusType::MOVEMENT)?;
-
-        Ok(Self {
-            chain_id: ChainId::new(chain_id.to_string()),
-            ibc_handler_address: config.ibc_handler_address,
-            aptos_client,
-            l1_settlement_address: config.l1_settlement_address,
-            l1_client_id: config.l1_client_id,
-            movement_rest_url: config.movement_rest_url,
-        })
-    }
-}
-
-impl aptos_move_ibc::ibc::ClientExt for Module {
-    fn client(&self) -> &aptos_rest_client::Client {
-        &self.aptos_client
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct Config {
-    /// The address of the `IBCHandler` smart contract.
-    pub ibc_handler_address: AccountAddress,
-
-    /// The address of the settlement contract on Eth.
-    pub l1_settlement_address: H160,
-
-    /// Id of the light client that this client depends on
-    pub l1_client_id: u32,
-
-    /// The RPC endpoint for aptos.
-    pub aptos_rest_api: String,
-
-    /// The RPC endpoint for custom movement apis.
-    pub movement_rest_url: String,
-}
-
-impl Module {
-    pub async fn ledger_version_of_height(&self, height: u64) -> u64 {
-        let ledger_version = self
-            .aptos_client
-            .get_block_by_height(height, false)
-            .await
-            // .map_err(rest_error_to_rpc_error)?
-            .unwrap()
-            .into_inner()
-            .last_version
-            .0;
-
-        debug!("height {height} is ledger version {ledger_version}");
-
-        ledger_version
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum ModuleInitError {
-    #[error("provider error")]
-    Rest(#[from] RestError),
-}
-
-#[async_trait]
-impl ConsensusModuleServer for Module {
-    /// Query the latest finalized height of this chain.
-    async fn query_latest_height(&self, _: &Extensions, _finalized: bool) -> RpcResult<Height> {
-        match self.aptos_client.get_index().await {
-            Ok(ledger_info) => {
-                let height = ledger_info.inner().block_height.0;
-
-                trace!(height, "latest height");
-
-                Ok(Height::new(height))
-            }
-            Err(err) => Err(ErrorObject::owned(
-                -1,
-                ErrorReporter(err).to_string(),
-                None::<()>,
-            )),
-        }
-    }
-
-    /// Query the latest finalized timestamp of this chain.
-    async fn query_latest_timestamp(
-        &self,
-        ext: &Extensions,
-        finalized: bool,
-    ) -> RpcResult<Timestamp> {
-        let latest_height = self.query_latest_height(ext, finalized).await?;
-
-        match self
-            .aptos_client
-            .get_block_by_height(latest_height.height(), false)
-            .await
-        {
-            Ok(block) => {
-                let timestamp = block.inner().block_timestamp.0;
-
-                debug!(%timestamp, %latest_height, "latest timestamp");
-
-                Ok(Timestamp::from_nanos(timestamp))
-            }
-            Err(err) => Err(ErrorObject::owned(
-                -1,
-                ErrorReporter(err).to_string(),
-                None::<()>,
-            )),
-        }
-    }
-}
