+++ voyager/modules/finality/berachain/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name    = "voyager-finality-module-berachain"
+version = "0.0.0"
+
+authors      = { workspace = true }
+edition      = { workspace = true }
+license-file = { workspace = true }
+publish      = { workspace = true }
+repository   = { workspace = true }
+
+[lints]
+workspace = true
+
+[dependencies]
+alloy                         = { workspace = true, features = ["rpc", "rpc-types", "transports", "transport-http", "transport-ws", "reqwest", "provider-ws"] }
+beacon-api-types              = { workspace = true, features = ["serde", "ssz"] }
+berachain-light-client-types  = { workspace = true, features = ["serde"] }
+cometbft-rpc                  = { workspace = true }
+dashmap                       = { workspace = true }
+embed-commit                  = { workspace = true }
+enumorph                      = { workspace = true }
+futures                       = { workspace = true }
+ics23                         = { workspace = true }
+jsonrpsee                     = { workspace = true, features = ["macros", "server", "tracing"] }
+macros                        = { workspace = true }
+num-bigint                    = { workspace = true }
+prost                         = { workspace = true }
+protos                        = { workspace = true }
+serde                         = { workspace = true, features = ["derive"] }
+serde_json                    = { workspace = true }
+tendermint-light-client-types = { workspace = true, features = ["proto", "serde"] }
+thiserror                     = { workspace = true }
+tokio                         = { workspace = true }
+tracing                       = { workspace = true }
+tracing-subscriber            = { workspace = true }
+unionlabs                     = { workspace = true, features = ["ssz"] }
+voyager-message               = { workspace = true }
+voyager-vm                    = { workspace = true }
