+++ voyager/modules/finality/bob/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name    = "voyager-finality-module-bob"
+version = "0.0.0"
+
+authors      = { workspace = true }
+edition      = { workspace = true }
+license-file = { workspace = true }
+publish      = { workspace = true }
+repository   = { workspace = true }
+
+[lints]
+workspace = true
+
+[dependencies]
+alloy                  = { workspace = true, features = ["rpc", "rpc-types", "transports", "transport-http", "transport-ws", "reqwest", "provider-ws"] }
+bob-client             = { workspace = true }
+bob-light-client-types = { workspace = true }
+bob-types              = { workspace = true }
+bob-verifier           = { workspace = true }
+embed-commit           = { workspace = true }
+ibc-union-spec         = { workspace = true, features = ["serde", "ethabi"] }
+jsonrpsee              = { workspace = true, features = ["macros", "server", "tracing"] }
+serde                  = { workspace = true, features = ["derive"] }
+tokio                  = { workspace = true }
+tracing                = { workspace = true }
+unionlabs              = { workspace = true }
+voyager-message        = { workspace = true }
+voyager-vm             = { workspace = true }
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["full"] }
