+++ voyager/modules/finality/ethereum/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name    = "voyager-finality-module-ethereum"
+version = "0.0.0"
+
+authors      = { workspace = true }
+edition      = { workspace = true }
+license-file = { workspace = true }
+publish      = { workspace = true }
+repository   = { workspace = true }
+
+[lints]
+workspace = true
+
+[dependencies]
+alloy            = { workspace = true, features = ["rpc", "rpc-types", "transports", "transport-http", "transport-ws", "reqwest", "provider-ws"] }
+beacon-api       = { workspace = true }
+beacon-api-types = { workspace = true, features = ["serde"] }
+embed-commit     = { workspace = true }
+jsonrpsee        = { workspace = true, features = ["macros", "server", "tracing"] }
+moka             = { version = "0.12.10", features = ["future"] }
+serde            = { workspace = true, features = ["derive"] }
+tokio            = { workspace = true }
+tracing          = { workspace = true }
+unionlabs        = { workspace = true }
+voyager-message  = { workspace = true }
+voyager-vm       = { workspace = true }
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["full"] }
